+++ src/MongoDBQueueServiceProvider.php
@@ -9,6 +9,9 @@ use Illuminate\Queue\QueueServiceProvider;
 use MongoDB\Laravel\Queue\Failed\MongoFailedJobProvider;
 
 use function array_key_exists;
+use function trigger_error;
+
+use const E_USER_DEPRECATED;
 
 class MongoDBQueueServiceProvider extends QueueServiceProvider
 {
@@ -51,6 +54,15 @@ class MongoDBQueueServiceProvider extends QueueServiceProvider
      */
     protected function mongoFailedJobProvider(array $config): MongoFailedJobProvider
     {
-        return new MongoFailedJobProvider($this->app['db'], $config['database'], $config['table']);
+        if (! isset($config['collection']) && isset($config['table'])) {
+            trigger_error('Since mongodb/laravel-mongodb 4.4: Using "table" option for the queue is deprecated. Use "collection" instead.', E_USER_DEPRECATED);
+            $config['collection'] = $config['table'];
+        }
+
+        return new MongoFailedJobProvider(
+            $this->app['db'],
+            $config['database'] ?? null,
+            $config['collection'] ?? 'failed_jobs',
+        );
     }
 }